backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML/AI
transformers==4.37.0
torch==2.1.2
spacy==3.7.2
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
Pillow==10.2.0

# Supabase
supabase==2.3.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# CORS
starlette==0.35.1
